Move report api spy to support folder
