+++ spec/cucumber/core_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'report_api_spy'
+require 'support/report_api_spy'
 require 'cucumber/core'
 require 'cucumber/core/filter'
 require 'cucumber/core/gherkin/writer'
