+++ spec/support/report_api_spy.rb
